Fix: Coordinates of player and supply should be real number

+++ src/agent/player.py
@@ -6,7 +6,7 @@ class Firearm:
         self.Distance = distance
 
 class Position:
-    def __init__(self, x: int, y: int):
+    def __init__(self, x: float, y: float):
         self.X = x
         self.Y = y
